requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
requests==2.32.3
beautifulsoup4==4.12.3
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.3
lxml==5.3.0

